next

+++ plugins/check_disk.c
@@ -995,7 +995,7 @@ get_stats (struct parameter_list *p, struct fs_usage *fsp) {
         get_fs_usage (p_list->best_match->me_mountdir, p_list->best_match->me_devname, &tmpfsp);
         get_path_stats(p_list, &tmpfsp);
         if (verbose >= 3)
-          printf("Group %s: adding %"PRIu64" blocks sized %"PRIu64", (%s) used_units=%g free_units=%g total_units=%g fsu_blocksize=%"PRIu64" mult=%"PRIu64"\n",
+          printf("Group %s: adding %lu blocks sized %"PRIu64", (%s) used_units=%g free_units=%g total_units=%g fsu_blocksize=%"PRIu64" mult=%"PRIu64"\n",
                  p_list->group, tmpfsp.fsu_bavail, tmpfsp.fsu_blocksize, p_list->best_match->me_mountdir, p_list->dused_units, p_list->dfree_units,
                  p_list->dtotal_units, mult);
 
@@ -1017,7 +1017,7 @@ get_stats (struct parameter_list *p, struct fs_usage *fsp) {
       }
       if (verbose >= 3)
         printf("Group %s now has: used_units=%g free_units=%g total_units=%g fsu_blocksize=%"PRIu64" mult=%"PRIu64"\n",
-               p->group, tmpfsp.fsu_bavail, tmpfsp.fsu_blocksize, p->best_match->me_mountdir, p->dused_units,
+               p->group, (uintmax_t)tmpfsp.fsu_bavail, tmpfsp.fsu_blocksize, p->best_match->me_mountdir, p->dused_units,
                p->dfree_units, p->dtotal_units, mult);
     }
     /* modify devname and mountdir for output */
